safari-600.7-branch/LayoutTests/fast/regions/cssom/webkit-named-flow-event-region-in-flexbox-overset-change-expected.txt

Tests the regionOversetChange event for regions inside flexboxes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Flowing content into regions...
PASS event.target.name is "article"
Engaging flexbox...
PASS event.target.name is "article"
PASS eventFiredCount is 2
PASS successfullyParsed is true

TEST COMPLETE
